PT-726 made order id as BigInteger

+++ client/src/main/kotlin/com/rarible/x2y2/client/model/Order.kt
@@ -8,7 +8,7 @@ import scalether.domain.Address
 
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy::class)
 data class Order(
-    val id: Int,
+    val id: BigInteger,
     val maker: Address,
     val currency: Address,
     val taker: Address?,
